% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{collect_software_versions}
\alias{collect_software_versions}
\title{Collect software versions}
\usage{
collect_software_versions(conda_envs, packages = TRUE)
}
\arguments{
\item{conda_envs}{Names of \code{conda} environments to query as a named list. The assumption is that the \code{conda} environment(s) actually exist!}

\item{packages}{Should versions of the currently loaded packages be included?}
}
\description{
Collect the version of software and packages loaded for the R session
}
\examples{
list(r=Sys.getenv('CONDA_DEFAULT_ENV'),
     project=read_dotbabs()$Lims) \%>\%
collect_software_versions(packages=FALSE)

}
