% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{ensembl_release_to_archive}
\alias{ensembl_release_to_archive}
\title{Retrieve Ensembl archive URL}
\usage{
ensembl_release_to_archive(release)
}
\arguments{
\item{release}{Ensembl release number, \code{character} or \code{integer}}
}
\description{
Get a URL to Ensembl archive server for a specific Ensemble release.
}
