% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{filter_dds_by_test_type}
\alias{filter_dds_by_test_type}
\title{Select \code{DESeqDataSet} objects from a list}
\usage{
filter_dds_by_test_type(x, type = "Wald")
}
\arguments{
\item{x}{A list of \code{DESeqDataSet} objects}

\item{type}{Either 'Wald' or 'LRT' to indicate which should be kept}
}
\value{
A list of \code{DESeqDataSet} objects; if x is a \code{DESeqDataSet} object (not a list) then a list is still returned of the one object, named \code{x}.
}
\description{
Filter a list of \code{DESeqDataSet} objects by the type of test that was run. Test type is determined by the content of the metadata of the object; existence of a \code{reduced} element indicates the test was \code{LRT}.
}
