% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cools.R
\name{get_cool_interaction_matrix}
\alias{get_cool_interaction_matrix}
\alias{get_cool_interaction_submatrix}
\title{Read data from a cool (or mcool) file}
\usage{
get_cool_interaction_matrix(cool, mcool, resolution)

get_cool_interaction_submatrix(
  chrom1,
  pos1,
  range1,
  chrom2 = chrom1,
  pos2 = pos1,
  range2 = range1,
  mcool,
  resolution
)
}
\arguments{
\item{cool, mcool}{Path to the h5 file}

\item{mcool, resolution}{Path to h5 file and resolution to extract}

\item{resolution}{Bin size in the \code{mcool} to extract}

\item{chrom1, chrom2}{Chromosomes for intervals}

\item{pos1, pos2}{Start coordinate of intervals}

\item{range1, range2}{Size of intervals (base pairs)}
}
\value{
A \code{data.frame} of 2D genomic coordinates with their weights and balanced score.
}
\description{
Reads the cool h5 file to provide the whole interaction matrix or a subset of coordinates.
}
\details{
The \code{cool} path takes priority and if \code{mcool} is provided so too must \code{resolution}.

Coordinates are filtered from the whole matrix via \code{get_cool_interaction_matrix()}.
}
\section{Functions}{
\itemize{
\item \code{get_cool_interaction_matrix}: Get a genome-wide interaction matrix

\item \code{get_cool_interaction_submatrix}: Get an interaction matrix for a pair of genomic regions
}}

