% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_ensembl_url}
\alias{get_ensembl_url}
\title{Format a URL to Ensembl}
\usage{
get_ensembl_url(ensembl_gene_id, mart)
}
\arguments{
\item{ensembl_gene_id}{Ensembl gene identifier for which a URL is created}

\item{mart}{The (optional) \code{biomaRt} connection object. If omitted a URL to search the current Ensembl release is produced.}
}
\description{
Uses \code{mart} object to produce a URL to a specific feature in Ensembl.
}
