% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_mart}
\alias{get_mart}
\title{Create a biomaRt connection}
\usage{
get_mart(species = "mmusculus", release = 95, dataset = "gene_ensembl", ...)
}
\arguments{
\item{species}{The organism name eg 'mmusulus'}

\item{release}{Ensembl release version}

\item{dataset}{Ensembl dataset to connect}
}
\description{
Connect to an Ensembl release server, providing species and release to get a \code{mart} object.
}
