% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{headtail}
\alias{headtail}
\title{Get the top and bottom of a \code{data.frame}}
\usage{
headtail(x, n = 10, n.head = n, n.tail = n)
}
\arguments{
\item{x}{A \code{data.frame}}

\item{n, n.head, n.tail}{Number of rows to head and tail}
}
\value{
A \code{data.frame} of \code{n.head} and \code{n.tail} rows.
}
\description{
Uses \code{rbind} to return the \code{head} and \code{tail} of a \code{data.frame}.
}
