% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatting.R
\name{iteratively_split_df}
\alias{iteratively_split_df}
\title{Split a data frame into a nested list}
\usage{
iteratively_split_df(
  x,
  variables = {
     head(colnames(x), n = -1)
 },
  f,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} to split}

\item{variables}{An ordered character vector of variables in \code{x} that will be nested levels of the output list}

\item{f}{A function applied to the data frame, once all variables have been used}

\item{...}{Arguments passed to \code{f}}
}
\value{
A nested list
}
\description{
Convert a set of data frame variables into keys of a nested list
}
\details{
\emph{May} assume that every combination of \code{variables} produces a data frame to which \code{f} can be applied.
}
