% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_caches.R
\name{lazyload_cached_chunk}
\alias{lazyload_cached_chunk}
\title{Lazy load a knitr chunk}
\usage{
lazyload_cached_chunk(
  chunk = "unnamed-chunk",
  filter,
  path = dirname(chunk),
  env = globalenv()
)
}
\arguments{
\item{chunk}{Name of cached chunk (from Rmd). Can be a path to a chunk (partial), from which `path` will be taken,}

\item{filter}{Character vector of object names to retain}

\item{path}{Root path within which caches (`*rdx`) are searched}

\item{env}{Environment into which objects are loaded}
}
\description{
Lazy load a knitr chunk
}
