% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cools.R
\name{list_mcool_resolutions}
\alias{list_mcool_resolutions}
\title{Get the available resolutions of an mcool}
\usage{
list_mcool_resolutions(mcool)
}
\arguments{
\item{mcool}{Path to the mcool}
}
\description{
Get the available resolutions of an mcool
}
