% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_caches.R
\name{load_all_chunks}
\alias{load_all_chunks}
\title{Lazy load all chunks in a cache}
\usage{
load_all_chunks(cache = "cache")
}
\arguments{
\item{cache}{Path to the cache from which all chunks are lazily loaded}
}
\description{
Lazy load all chunks in a cache
}
