% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_caches.R
\name{load_cached_chunk}
\alias{load_cached_chunk}
\alias{load_all_chunks}
\title{Load a cached knitr chunk}
\usage{
load_cached_chunk(
  chunk = "unnamed-chunk",
  cache = c("knitr_cache", "cache"),
  filter,
  env = globalenv()
)

load_all_chunks(cache = c("knitr_cache", "cache"))
}
\arguments{
\item{chunk}{Name of cached chunk (from Rmd). Can be a path to a chunk (partial), from which \code{path} will be taken.}

\item{cache}{Path to the cache from which all chunks are loaded}

\item{filter}{Character vector of object names to retain}

\item{env}{Environment into which objects are loaded}
}
\description{
Loads the cached chunk from cache using lazy loading where possible.
}
\details{
Any existing objects in the environment will be overwritten when chunk(s) are loaded.
}
\section{Functions}{
\itemize{
\item \code{load_cached_chunk()}: Load a cached chunk

\item \code{load_all_chunks()}: Load all cached chunks (alphabetically)

importFrom plyr l_ply

}}
