% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_scales.R
\name{log_breaks}
\alias{log_breaks}
\alias{major_log_breaks}
\alias{minor_log_breaks}
\title{Get breaks for a log scale}
\usage{
log_breaks(base = 10, max.exp = 7, min.exp = 0)

major_log_breaks(base = 10, max.exp = 7, min.exp = 0)

minor_log_breaks(base = 10, max.exp = 7, min.exp = 0)
}
\arguments{
\item{base}{Log base}

\item{max.exp, min.exp}{range of exponents to return}
}
\description{
Provides breaks for major and minor log-transformed scales.
}
\section{Functions}{
\itemize{
\item \code{log_breaks}: Log breaks

\item \code{major_log_breaks}: Major log breaks

\item \code{minor_log_breaks}: Minor log breaks
}}

\examples{
log_breaks(base=10, max.exp=2)

major_log_breaks(base=10, max.exp=2)

minor_log_breaks(base=10, max.exp=2)

}
\concept{log breaks}
