% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{make_clean_directory}
\alias{make_clean_directory}
\title{Create and/or clear out a directory}
\usage{
make_clean_directory(path, clean = FALSE)
}
\arguments{
\item{path}{Path to directory to create and/or clean}

\item{clean}{Should the contents of \code{path} be deleted? Defaults to \code{FALSE} for safety.}
}
\value{
Invisibly returns the \code{path} that may have been created and/or cleaned.
}
\description{
Create and/or clear out a directory
}
