% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatting.R
\name{numeric_to_si}
\alias{numeric_to_si}
\title{Add metric suffixes to numbers}
\usage{
numeric_to_si(i, sep = "")
}
\arguments{
\item{i}{Number to convert (passed through \code{as.numeric})}

\item{sep}{Separator between number and suffix}
}
\value{
Character vector of equal size to \code{i}.
}
\description{
Convert numeric values into characters with metric suffixes.
}
