% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{open_x11}
\alias{open_x11}
\title{Open X11 on a display}
\usage{
open_x11(display_id = 10, max_attempts = 10)
}
\arguments{
\item{display_id}{Starting display ID}

\item{max_attempts}{Number of displays to try}
}
\value{
Invisibly returns the opened display ID.
}
\description{
Iterate through displays to find an X11 display to use.
}
\details{
When reconnecting a \code{tmux} session, the R session may no longer be able to open an X11 window. In such cases, the display ID may have changed between the R session opening and the \code{tmux} session reconnecting. This function iterates over the displays until one can be used or \code{max_attempts} is reached.
}
