% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{preferred_choice}
\alias{preferred_choice}
\title{Make a ranked choice}
\usage{
preferred_choice(x, preferences, default = 1, missing = NA)
}
\arguments{
\item{x}{Named vector of possible choices}

\item{preferences}{Ranked vector of names of \code{x}}

\item{default}{Element of \code{x} if none of \code{preferences} are names of \code{x}}

\item{missing}{Value returned instead of \code{x[default]}, when `default=0` or none of \code{preferences} or \code{default} exist in \code{names(x)}}
}
\description{
Select a vector element using a ranked set of preferences.
}
