% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{print_object_size}
\alias{print_object_size}
\title{Get the size of an object}
\usage{
print_object_size(x, unit = "GB")
}
\arguments{
\item{x}{Object to check}

\item{unit}{Units to print}
}
\description{
Prints the size in \code{units} of an object(s).
}
\seealso{
utils:::format.object_size
}
