% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babs.R
\name{read_msigdb}
\alias{read_msigdb}
\title{Read files from my MSigDB cache}
\usage{
read_msigdb(
  collection = "h.all",
  pathways = NULL,
  version = "7.4",
  identifier = "symbols",
  dbroot = "/camp/stp/babs/working/barrinc/db/msigdb"
)
}
\arguments{
\item{collection}{Name of the collection}

\item{pathways}{Character vector of pathways in \code{collection} to keep}

\item{version}{Release version}

\item{identifier}{Identifier type: \code{symbols} or \code{entrez} (\code{entrez} may not exist!!)}

\item{dbroot}{Path to the root of the MSigDB files}
}
\value{
A list of genes in pathways, with pathway name as the key and a character vector of gene identifiers (\code{identifer}).
}
\description{
Reads an MSigDB \code{gmt} file and returns a list of genes in pathways.
}
