% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_panels.R
\name{remove_clipping}
\alias{remove_clipping}
\title{Prevent clipping in \code{grid} objects}
\usage{
remove_clipping(x)
}
\arguments{
\item{x}{Grid object}
}
\description{
Prevent clipping in \code{grid} objects
}
