% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_panels.R
\name{resize_and_show}
\alias{resize_and_show}
\title{Render a \code{ggplot} as a \code{gtable}}
\usage{
resize_and_show(
  x,
  size,
  width,
  height,
  unit = "in",
  orientation = c("landscape", "portrait"),
  aspect = 1.6
)
}
\arguments{
\item{x}{Either a \code{ggplot} or \code{gtable}/\code{pheatmap}}

\item{size}{Value of both \code{width} and \code{height}, if defined}

\item{width, height}{Numeric values for size of the dimensions. Set to \code{NULL} to avoid resizing.}

\item{unit}{Character for units of \code{width} and \code{height}, passed to \code{\link[grid:unit]{grid::unit()}}}

\item{orientation}{Character either 'landscape'/'l' or 'portrait'/'p'}

\item{aspect}{Numeric for aspect ratio to apply}
}
\description{
Accepts a \code{ggplot} object and converts it to a \code{gtable} before setting panel dimensions and returning (and showing) the modified \code{gtable}.
}
\seealso{
\code{\link[grid:unit]{grid::unit()}}
}
