% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_panels.R
\name{set_panel_dims}
\alias{set_panel_dims}
\title{Define the dimensions of a \code{gtable} panel}
\usage{
set_panel_dims(
  ggplot = NULL,
  ggplot_gtable = ggplotGrob(ggplot),
  height = width,
  width = height
)
}
\arguments{
\item{ggplot}{Input \code{ggplot} object}

\item{ggplot_gtable}{Input \code{gtable} object}

\item{height, width}{Dimensions in \code{\link[grid:unit]{grid::unit()}}}
}
\description{
Define the dimensions of a \code{gtable} panel
}
