% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_panels.R
\name{show_resized_plot}
\alias{show_resized_plot}
\title{Print a \code{grid} object}
\usage{
show_resized_plot(x)
}
\arguments{
\item{x}{Object to display on a new page}
}
\description{
Print a \code{grid} object
}
