% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{wideScreen}
\alias{wideScreen}
\alias{get_screen_width}
\title{Set width of console}
\usage{
wideScreen(how_wide = get_screen_width() - 1, verbose = FALSE)

get_screen_width()
}
\arguments{
\item{how_wide}{Number of characters}

\item{verbose}{Include message of new console width}
}
\description{
Uses the width of the console to define the number of characters to print on lines.
}
\section{Functions}{
\itemize{
\item \code{wideScreen()}: Set width of the console output

\item \code{get_screen_width()}: Get width of console

}}
