% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{write_parameters_file}
\alias{write_parameters_file}
\title{Write parameters file}
\usage{
write_parameters_file(x, filename = "parameters.csv", add_id = TRUE)
}
\arguments{
\item{x}{\code{data.frame} of parameters}

\item{filename}{Path to output file}

\item{add_id}{Should an \code{id} variable be included, which is just a zero-padded row number}
}
\description{
Writes a \code{data.frame} as a \code{csv} ready to be ready by an \code{sbatch} job array.
}
