% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Get the extension of a file}
\usage{
get_file_extension(path)
}
\arguments{
\item{path}{At least a file name, potentially the full path}
}
\description{
Get the extension of a file
}
