% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaying_data.R
\name{kable}
\alias{kable}
\title{Set default options for a kable}
\usage{
kable(...)
}
\arguments{
\item{...}{Options passed to \code{kableExtra::kable}}
}
\description{
Uses options that I always have to type manually.
}
