% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo.R
\name{make_tabs_shortcode}
\alias{make_tabs_shortcode}
\title{Make tabs shortcode}
\usage{
make_tabs_shortcode(x, tabset_name = "chunk_name")
}
\arguments{
\item{x}{List (of lists) from which tabs will be created}
}
\value{
A character vector for the shortcode to include in the markdown.
}
\description{
Write a shortcode to create tabs from a (nested) list.
}
