% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\docType{data}
\name{project_paths}
\alias{project_paths}
\title{Paths for the project}
\format{
An object of class \code{list} of length 8.
}
\usage{
project_paths
}
\arguments{
\item{...}{Ignored.}
}
\value{
Each function returns a character string absolute path.
}
\description{
A list of functions that provide paths to different parts of the project.
}
\keyword{datasets}
