% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_panels.R
\name{resize_and_show}
\alias{resize_and_show}
\title{Render a ggplot as a gtable}
\usage{
resize_and_show(x, width = height * 1.6, height = width/1.6, unit = "in")
}
\arguments{
\item{x}{Either a \code{ggplot} or \code{gtable}}

\item{width, height}{Numeric values for size of the dimensions}

\item{unit}{Character for units of \code{width} and \code{height}, passed to \code{grid::unit}}
}
\description{
Accepts a \code{ggplot} object and converts it to a \code{gtable} before setting panel dimensions and returning (and showing) the modified \code{gtable}.
}
\seealso{
grid::unit
}
